requirements.txt
django
djangorestframework
psycopg2-binary
django-countries
djangorestframework-simplejwt
django-filter
drf-nested-routers
pillow